+++ matholymp/csvsource.py
@@ -244,6 +244,7 @@ class CSVDataSource(DataSource):
                                   'departure_place': 'Departure Place',
                                   'departure_flight': 'Departure Flight',
                                   'consent_form_url': 'Consent Form URL',
+                                  'id_scan_url': 'ID Scan URL',
                                   'passport_number':
                                   'Passport or Identity Card Number',
                                   'nationality': 'Nationality',
@@ -361,6 +362,20 @@ class CSVDataSource(DataSource):
                 return file_url_to_local(url, os.path.join(self._local_dir,
                                                            'consent-forms'),
                                          'consent-form', person_id)
+            else:
+                # Not actually a possible case.
+                url_path = url[len(self._cfg['url_base']):]
+                url_dirs = url_path.split('/')
+                return os.path.join(self._local_dir, *url_dirs)
+        if name == 'id_scan_filename':
+            k = 'ID Scan URL'
+            url = self._people[event_id][person_id][country_id][k]
+            if url == '':
+                return None
+            if self._reg_system:
+                return file_url_to_local(url, os.path.join(self._local_dir,
+                                                           'id-scans'),
+                                         'id-scan', person_id)
             else:
                 # Not actually a possible case.
                 url_path = url[len(self._cfg['url_base']):]