+++ matholymp/roundupreg/schema.py
@@ -55,7 +55,7 @@
 """This module provides the Roundup registration schema."""
 
 from matholymp.roundupreg.config import distinguish_official, \
-    have_consent_forms, have_consent_ui, have_passport_numbers, \
+    have_consent_forms, have_id_scans, have_consent_ui, have_passport_numbers, \
     have_nationality, get_language_numbers, invitation_letter_register, \
     is_virtual_event, have_remote_participation, allow_hybrid_countries, \
     get_sars_cov2_cert_bool, get_sars_cov2_doses_bool, \
@@ -171,6 +171,8 @@ def init_schema(env):
     person_extra = {}
     if have_consent_forms(db):
         person_extra['consent_form'] = Link('consent_form')
+    if have_id_scans(db):
+        person_extra['id_scan'] = Link('id_scan')
     if have_consent_ui(db):
         person_extra['event_photos_consent'] = Boolean()
         # not_applicable, no, badge_only, yes.
@@ -265,6 +267,10 @@ def init_schema(env):
               name=String(), person=Link('person'))
 
     FileClass(db, 'consent_form',
+              content=String(indexme='no'),
+              name=String(), person=Link('person'))
+
+    FileClass(db, 'id_scan',
               content=String(indexme='no'),
               name=String(), person=Link('person'))
 
@@ -382,6 +388,8 @@ def init_schema(env):
                                'photo', 'incomplete']
     if have_consent_forms(db):
         person_common_reg_props.append('consent_form')
+    if have_id_scans(db):
+        person_common_reg_props.append('id_scan')
     if have_consent_ui(db):
         person_common_reg_props.append('event_photos_consent')
         person_common_reg_props.append('photo_consent')
@@ -488,6 +496,41 @@ def init_schema(env):
     p = db.security.addPermission(name='View', klass='photo',
                                   check=own_person_photo)
     db.security.addPermissionToRole('SelfRegister', p)
+
+    # Registering users can create ID scans, and view them only
+    # when from their own country or created by that user (the latter
+    # as a case for access before the person is set).
+    db.security.addPermissionToRole('Register', 'Create', 'id_scan')
+    # Self-registering users can create and see their own ID scans.
+    db.security.addPermissionToRole('SelfRegister', 'Create', 'id_scan')
+
+    def own_country_id_scan(db, userid, itemid):
+        """Determine whether the userid matches the country of the ID 
+        scan being accessed."""
+        user_country = db.user.get(userid, 'country')
+        file_person = db.id_scan.get(itemid, 'person')
+        file_country = None
+        if file_person:
+            file_country = db.person.get(file_person, 'country')
+        file_creator = db.id_scan.get(itemid, 'creator')
+        return (user_country == file_country
+                or (file_country is None and userid == file_creator))
+
+    def own_person_id_scan(db, userid, itemid):
+        """Determine whether the userid matches the person of the ID 
+        scan being accessed."""
+        user_person = db.user.get(userid, 'person')
+        file_person = db.id_scan.get(itemid, 'person')
+        file_creator = db.id_scan.get(itemid, 'creator')
+        return (user_person == file_person
+                or (file_person is None and userid == file_creator))
+
+    p = db.security.addPermission(name='View', klass='id_scan',
+                                  check=own_country_id_scan)
+    db.security.addPermissionToRole('Register', p)
+    p = db.security.addPermission(name='View', klass='id_scan',
+                                  check=own_person_id_scan)
+    db.security.addPermissionToRole('Register', p)
 
     # Registering users can create consent forms, and view them only
     # when from their own country or created by that user (the latter